Add files via upload
The 911 Calls Capstone Project is an exploratory data analysis (EDA) project that uses a dataset containing information about emergency calls made to the 911 service. The project involves analyzing the data to uncover insights and trends related to the calls made, such as the time of day, day of the week, and type of emergency. Seaborn, a powerful plotting library, is employed to create visualizations that help to identify patterns and relationships within the data.