Add files via upload
This data project involves an exploratory analysis of stock prices, with a focus on bank stocks and their performance during the financial crisis of 2007-2008, up until early 2016. The project is designed to provide practice in using visualization and pandas skills, and is not intended to be a comprehensive financial analysis or provide financial advice. By analyzing and visualizing the data, we aim to gain insights into the trends and patterns that emerged during this period, as well as the impact of the financial crisis on bank stocks. This project offers an opportunity to develop and refine data analysis skills, while also gaining a deeper understanding of the dynamics of the financial market.